Fixed numerous compiler warnings

+++ libtwolame/get_set.c
@@ -100,7 +100,7 @@ TL_API int twolame_set_scale(twolame_options * glopts, float scale)
 
 TL_API float twolame_get_scale(twolame_options * glopts)
 {
-    return (glopts->scale);
+    return (float)(glopts->scale);
 }
 
 TL_API int twolame_set_scale_left(twolame_options * glopts, float scale)
@@ -115,7 +115,7 @@ TL_API int twolame_set_scale_left(twolame_options * glopts, float scale)
 
 TL_API float twolame_get_scale_left(twolame_options * glopts)
 {
-    return (glopts->scale_left);
+    return (float)(glopts->scale_left);
 }
 
 TL_API int twolame_set_scale_right(twolame_options * glopts, float scale)
@@ -130,7 +130,7 @@ TL_API int twolame_set_scale_right(twolame_options * glopts, float scale)
 
 TL_API float twolame_get_scale_right(twolame_options * glopts)
 {
-    return (glopts->scale_right);
+    return (float)(glopts->scale_right);
 }
 
 
@@ -292,7 +292,7 @@ TL_API int twolame_set_VBR_level(twolame_options * glopts, float level)
 
 TL_API float twolame_get_VBR_level(twolame_options * glopts)
 {
-    return (glopts->vbrlevel);
+    return (float)(glopts->vbrlevel);
 }
 
 TL_API int twolame_set_ATH_level(twolame_options * glopts, float level)
@@ -303,7 +303,7 @@ TL_API int twolame_set_ATH_level(twolame_options * glopts, float level)
 
 TL_API float twolame_get_ATH_level(twolame_options * glopts)
 {
-    return (glopts->athlevel);
+    return (float)(glopts->athlevel);
 }
 
 TL_API int twolame_set_quick_mode(twolame_options * glopts, int quickmode)
